+++ IRasRag.Domain/Entities/FeedType.cs
@@ -13,7 +13,7 @@ namespace IRasRag.Domain.Entities
         public string Description { get; set; }
 
         [Required]
-        public float ProteinPercentage { get; set; } // % of protein
+        public double ProteinPercentage { get; set; } // % of protein
 
         [MaxLength(255)]
         public string Manufacturer { get; set; }
